project-boilerplates/sending-images-streamlit-fastapi/frontend/requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
six==1.14.0

# img
pillow

# web
streamlit
requests

# img
opencv-python
